Move functions from utils.py to utils package.
- The constants are better off as a separate file

# Conflicts:
#	rsmtool/utils.py
