+++ rsmtool/utils/constants.py
@@ -1,0 +1,195 @@
+"""
+Various RSMTool constants used across the codebase.
+
+:author: Jeremy Biggs (jbiggs@ets.org)
+:author: Anastassia Loukina (aloukina@ets.org)
+:author: Nitin Madnani (nmadnani@ets.org)
+
+:organization: ETS
+"""
+
+DEFAULTS = {'id_column': 'spkitemid',
+            'description': '',
+            'description_old': '',
+            'description_new': '',
+            'train_label_column': 'sc1',
+            'test_label_column': 'sc1',
+            'human_score_column': 'sc1',
+            'exclude_zero_scores': True,
+            'use_scaled_predictions': False,
+            'use_scaled_predictions_old': False,
+            'use_scaled_predictions_new': False,
+            'select_transformations': False,
+            'standardize_features': True,
+            'use_thumbnails': False,
+            'use_truncation_thresholds': False,
+            'scale_with': None,
+            'predict_expected_scores': False,
+            'sign': None,
+            'features': None,
+            'length_column': None,
+            'second_human_score_column': None,
+            'file_format': 'csv',
+            'form_level_scores': None,
+            'candidate_column': None,
+            'general_sections': ['all'],
+            'special_sections': None,
+            'custom_sections': None,
+            'feature_subset_file': None,
+            'feature_subset': None,
+            'trim_min': None,
+            'trim_max': None,
+            'trim_tolerance': 0.4998,
+            'subgroups': [],
+            'min_n_per_group': None,
+            'skll_objective': None,
+            'section_order': None,
+            'flag_column': None,
+            'flag_column_test': None,
+            'min_items_per_candidate': None,
+            'experiment_names': None}
+
+LIST_FIELDS = ['general_sections',
+               'special_sections',
+               'custom_sections',
+               'subgroups',
+               'section_order',
+               'experiment_dirs',
+               'experiment_names']
+
+BOOLEAN_FIELDS = ['exclude_zero_scores',
+                  'predict_expected_scores',
+                  'use_scaled_predictions',
+                  'use_scaled_predictions_old',
+                  'use_scaled_predictions_new',
+                  'use_thumbnails',
+                  'use_truncation_thresholds',
+                  'select_transformations']
+
+FIELD_NAME_MAPPING = {'expID': 'experiment_id',
+                      'LRmodel': 'model',
+                      'train': 'train_file',
+                      'test': 'test_file',
+                      'predictions': 'predictions_file',
+                      'feature': 'features',
+                      'train.lab': 'train_label_column',
+                      'test.lab': 'test_label_column',
+                      'trim.min': 'trim_min',
+                      'trim.max': 'trim_max',
+                      'scale': 'use_scaled_predictions',
+                      'feature.subset': 'feature_subset'}
+
+MODEL_NAME_MAPPING = {'empWt': 'LinearRegression',
+                      'eqWt': 'EqualWeightsLR',
+                      'empWtBalanced': 'RebalancedLR',
+                      'empWtDropNeg': '',
+                      'empWtNNLS': 'NNLR',
+                      'empWtDropNegLasso': 'LassoFixedLambdaThenNNLR',
+                      'empWtLasso': 'LassoFixedLambdaThenLR',
+                      'empWtLassoBest': 'PositiveLassoCVThenLR',
+                      'lassoWtLasso': 'LassoFixedLambda',
+                      'lassoWtLassoBest': 'PositiveLassoCV'}
+
+CHECK_FIELDS = {'rsmtool': {'required': ['experiment_id',
+                                         'model',
+                                         'train_file',
+                                         'test_file'],
+                            'optional': ['description',
+                                         'features',
+                                         'feature_subset_file',
+                                         'feature_subset',
+                                         'file_format',
+                                         'sign',
+                                         'id_column',
+                                         'use_thumbnails',
+                                         'train_label_column',
+                                         'test_label_column',
+                                         'length_column',
+                                         'second_human_score_column',
+                                         'flag_column',
+                                         'flag_column_test',
+                                         'exclude_zero_scores',
+                                         'trim_min',
+                                         'trim_max',
+                                         'trim_tolerance',
+                                         'predict_expected_scores',
+                                         'select_transformations',
+                                         'use_scaled_predictions',
+                                         'use_truncation_thresholds',
+                                         'subgroups',
+                                         'min_n_per_group',
+                                         'general_sections',
+                                         'custom_sections',
+                                         'special_sections',
+                                         'skll_objective',
+                                         'section_order',
+                                         'candidate_column',
+                                         'standardize_features',
+                                         'min_items_per_candidate']},
+                'rsmeval': {'required': ['experiment_id',
+                                         'predictions_file',
+                                         'system_score_column',
+                                         'trim_min',
+                                         'trim_max'],
+                            'optional': ['description',
+                                         'id_column',
+                                         'human_score_column',
+                                         'second_human_score_column',
+                                         'file_format',
+                                         'flag_column',
+                                         'exclude_zero_scores',
+                                         'use_thumbnails',
+                                         'scale_with',
+                                         'trim_tolerance',
+                                         'subgroups',
+                                         'min_n_per_group',
+                                         'general_sections',
+                                         'custom_sections',
+                                         'special_sections',
+                                         'section_order',
+                                         'candidate_column',
+                                         'min_items_per_candidate']},
+                'rsmpredict': {'required': ['experiment_id',
+                                            'experiment_dir',
+                                            'input_features_file'],
+                               'optional': ['id_column',
+                                            'candidate_column',
+                                            'file_format',
+                                            'predict_expected_scores',
+                                            'human_score_column',
+                                            'second_human_score_column',
+                                            'standardize_features',
+                                            'subgroups',
+                                            'flag_column']},
+                'rsmcompare': {'required': ['comparison_id',
+                                            'experiment_id_old',
+                                            'experiment_dir_old',
+                                            'experiment_id_new',
+                                            'experiment_dir_new',
+                                            'description_old',
+                                            'description_new'],
+                               'optional': ['use_scaled_predictions_old',
+                                            'use_scaled_predictions_new',
+                                            'subgroups',
+                                            'use_thumbnails',
+                                            'general_sections',
+                                            'custom_sections',
+                                            'special_sections',
+                                            'section_order']},
+                'rsmsummarize': {'required': ['summary_id',
+                                              'experiment_dirs'],
+                                 'optional': ['description',
+                                              'experiment_names',
+                                              'file_format',
+                                              'general_sections',
+                                              'custom_sections',
+                                              'use_thumbnails',
+                                              'special_sections',
+                                              'subgroups',
+                                              'section_order']}}
+
+ID_FIELDS = {'rsmtool': 'experiment_id',
+             'rsmeval': 'experiment_id',
+             'rsmcompare': 'comparison_id',
+             'rsmsummarize': 'summary_id',
+             'rsmpredict': 'experiment_id'}
