+++ rsmtool/utils/logging.py
@@ -1,0 +1,71 @@
+"""
+Utility classes and functions for logging.
+
+:author: Jeremy Biggs (jbiggs@ets.org)
+:author: Anastassia Loukina (aloukina@ets.org)
+:author: Nitin Madnani (nmadnani@ets.org)
+
+:organization: ETS
+"""
+
+import logging
+
+
+class LogFormatter(logging.Formatter):
+    """
+    Custom logging formatter.
+
+    Adapted from:
+        http://stackoverflow.com/questions/1343227/
+        can-pythons-logging-format-be-modified-depending-
+        on-the-message-log-level
+    """
+
+    info_fmt = "%(msg)s"
+    warn_fmt = "WARNING: %(msg)s"
+
+    err_fmt = "ERROR: %(msg)s"
+    dbg_fmt = "DEBUG: %(module)s: %(lineno)d: %(msg)s"
+
+    def __init__(self, fmt="%(levelno)s: %(msg)s"):
+
+        logging.Formatter.__init__(self, fmt)
+
+    def format(self, record):
+        """
+        format the logger
+
+        Parameters
+        ----------
+        record
+            The record to format
+        """
+
+        # Save the original format configured by the user
+        # when the logger formatter was instantiated
+        format_orig = self._fmt
+
+        # Replace the original format with one customized by logging level
+        if record.levelno == logging.DEBUG:
+            self._fmt = LogFormatter.dbg_fmt
+            self._style = logging.PercentStyle(self._fmt)
+
+        elif record.levelno == logging.WARNING:
+            self._fmt = LogFormatter.warn_fmt
+            self._style = logging.PercentStyle(self._fmt)
+
+        elif record.levelno == logging.INFO:
+            self._fmt = LogFormatter.info_fmt
+            self._style = logging.PercentStyle(self._fmt)
+
+        elif record.levelno == logging.ERROR:
+            self._fmt = LogFormatter.err_fmt
+            self._style = logging.PercentStyle(self._fmt)
+
+        # Call the original formatter class to do the grunt work
+        result = logging.Formatter.format(self, record)
+
+        # Restore the original format configured by the user
+        self._fmt = format_orig
+
+        return result
